make useEdgeClient handle http auth
