+++ packages/sdk/client/package.json
@@ -25,6 +25,11 @@
       "browser": "./dist/lib/browser/halo/index.mjs",
       "node": "./dist/lib/node-esm/halo/index.mjs"
     },
+    "./edge": {
+      "types": "./dist/types/src/edge/index.d.ts",
+      "browser": "./dist/lib/browser/edge/index.mjs",
+      "node": "./dist/lib/node-esm/edge/index.mjs"
+    },
     ".": {
       "types": "./dist/types/src/index.d.ts",
       "browser": "./dist/lib/browser/index.mjs",
@@ -63,6 +68,9 @@
       "halo": [
         "dist/types/src/halo/index.d.ts"
       ],
+      "edge": [
+        "dist/types/src/edge/index.d.ts"
+      ],
       "invitations": [
         "dist/types/src/invitations/index.d.ts"
       ],
@@ -96,6 +104,7 @@
     "@dxos/echo-pipeline": "workspace:*",
     "@dxos/echo-protocol": "workspace:*",
     "@dxos/echo-schema": "workspace:*",
+    "@dxos/edge-client": "workspace:*",
     "@dxos/indexing": "workspace:*",
     "@dxos/invariant": "workspace:*",
     "@dxos/keys": "workspace:*",
@@ -122,7 +131,6 @@
   },
   "devDependencies": {
     "@dxos/automerge": "workspace:*",
-    "@dxos/edge-client": "workspace:*",
     "@dxos/test-utils": "workspace:*",
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.isequal": "^4.5.0",
