+++ packages/sdk/client/project.json
@@ -13,6 +13,7 @@
         "entryPoints": [
           "{projectRoot}/src/devtools/index.ts",
           "{projectRoot}/src/echo/index.ts",
+          "{projectRoot}/src/edge/index.ts",
           "{projectRoot}/src/halo/index.ts",
           "{projectRoot}/src/index.ts",
           "{projectRoot}/src/invitations/index.ts",
