+++ packages/sdk/client/src/edge/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export * from './edge-identity';
