+++ packages/sdk/react-edge-client/package.json
@@ -20,6 +20,7 @@
     "src"
   ],
   "dependencies": {
+    "@dxos/client": "workspace:*",
     "@dxos/echo-db": "workspace:*",
     "@dxos/echo-signals": "workspace:*",
     "@dxos/edge-client": "workspace:*",
