+++ packages/sdk/react-edge-client/src/hooks/useEdgeClient.ts
@@ -4,9 +4,11 @@
 
 import { useMemo } from 'react';
 
+import { createEdgeIdentity } from '@dxos/client/edge';
 import { EdgeHttpClient } from '@dxos/edge-client';
 import { invariant } from '@dxos/invariant';
-import { useConfig } from '@dxos/react-client';
+import { useConfig, useClient } from '@dxos/react-client';
+import { useIdentity } from '@dxos/react-client/halo';
 
 /**
  * Client for edge services.
@@ -14,7 +16,14 @@ import { useConfig } from '@dxos/react-client';
 // TODO(mykola): This should be done through client-service RPC.
 export const useEdgeClient = () => {
   const config = useConfig();
+  const client = useClient();
+  const identity = useIdentity();
   const edgeUrl = config.values.runtime?.services?.edge?.url;
   invariant(edgeUrl, 'EDGE services not configured.');
-  return useMemo(() => new EdgeHttpClient(edgeUrl), [edgeUrl]);
+  return useMemo(() => {
+    const edgeClient = new EdgeHttpClient(edgeUrl);
+    const edgeIdentity = createEdgeIdentity(client);
+    edgeClient.setIdentity(edgeIdentity);
+    return edgeClient;
+  }, [edgeUrl, identity]);
 };
