+++ packages/sdk/react-edge-client/tsconfig.json
@@ -30,6 +30,9 @@
       "path": "../../core/mesh/edge-client"
     },
     {
+      "path": "../client"
+    },
+    {
       "path": "../react-client"
     }
   ]
